feat: show preview table
