+++ app/views/Data/Workspace/ImportTable/styles.module.css
@@ -1,0 +1,61 @@
+.import {
+    display: flex;
+    flex-direction: column;
+    padding: 0 var(--mantine-spacing-md);
+
+    .heading {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        padding-top: var(--mantine-spacing-sm);
+        padding-bottom: var(--mantine-spacing-sm);
+    }
+
+    .collapse {
+        display: flex;
+        flex-direction: column;
+
+        .content {
+            display: flex;
+            gap: var(--mantine-spacing-sm);
+
+            .table-properties-container {
+                display: flex;
+                flex-basis: 20%;
+                flex-direction: column;
+                gap: var(--mantine-spacing-sm);
+                background-color: var(--mantine-color-gray-1);
+                padding: var(--mantine-spacing-md) var(--mantine-spacing-sm);
+
+                .form {
+                    display: flex;
+                    flex-direction: column;
+                    gap: var(--mantine-spacing-md);
+
+                    .icon {
+                        color: var(--mantine-color-gray-5);
+                    }
+                }
+            }
+
+            .table-container {
+                display: flex;
+                position: relative;
+                flex-basis: 80%;
+                justify-content: center;
+                max-height: 45vh;
+                overflow: auto;
+
+                .alert {
+                    display: flex;
+                    align-items: center;
+                }
+            }
+        }
+
+        .import-actions {
+            display: flex;
+            padding: var(--mantine-spacing-sm) 0;
+        }
+    }
+}
