+++ app/views/Data/Workspace/styles.module.css
@@ -7,57 +7,4 @@
         display: flex;
         flex-grow: 1;
     }
-
-    .import {
-        display: flex;
-        flex-direction: column;
-        padding: 0 var(--mantine-spacing-md);
-
-        .heading {
-            display: flex;
-            align-items: center;
-            justify-content: space-between;
-            padding-top: var(--mantine-spacing-sm);
-            padding-bottom: var(--mantine-spacing-sm);
-        }
-
-        .collapse {
-            display: flex;
-            flex-direction: column;
-
-            .content {
-                display: flex;
-                gap: var(--mantine-spacing-sm);
-
-                .table-properties-container {
-                    display: flex;
-                    flex-basis: 20%;
-                    flex-direction: column;
-                    gap: var(--mantine-spacing-sm);
-                    background-color: var(--mantine-color-gray-1);
-                    padding: var(--mantine-spacing-md) var(--mantine-spacing-sm);
-
-                    .form {
-                        display: flex;
-                        flex-direction: column;
-                        gap: var(--mantine-spacing-md);
-
-                        .icon {
-                            color: var(--mantine-color-gray-5);
-                        }
-                    }
-                }
-
-                .table-container {
-                    display: flex;
-                    flex-basis: 80%;
-                }
-            }
-
-            .import-actions {
-                display: flex;
-                padding: var(--mantine-spacing-sm) 0;
-            }
-        }
-    }
 }
